Minor updates
Minor cleanup to the document.

Add Sphere.addDetector() and related support.
This is the first version of the detector class to do something useful.
This change is what I really wanted to commit.

This code is not attached to the GUI yet.
You can see it from the console by typing:
  s = spheres[2]
  d = s.addDetector()
It is set up like this for testing and experimenting.